+++ components/campaigns/const-mock.ts
@@ -1,30 +1,12 @@
+import { CampaignEventContition } from "@/app/api/generated/prisma";
 
-// TODO: Potentially fetch timezone list dynamically
-export const timezones = [
-  "(GMT-12:00) International Date Line West",
-  "(GMT-11:00) Midway Island, Samoa",
-  "(GMT-10:00) Hawaii",
-  "(GMT-09:00) Alaska",
-  "(GMT-08:00) Pacific Time (US & Canada)",
-  "(GMT-07:00) Mountain Time (US & Canada)",
-  "(GMT-06:00) Central Time (US & Canada)",
-  "(GMT-05:00) Eastern Time (US & Canada)",
-  "(GMT-04:00) Atlantic Time (Canada)",
-  "(GMT-03:00) Buenos Aires, Georgetown",
-  "(GMT-02:00) Mid-Atlantic",
-  "(GMT-01:00) Azores",
-  "(GMT+00:00) Greenwich Mean Time : Dublin, Edinburgh, Lisbon, London",
-  "(GMT+01:00) Amsterdam, Berlin, Bern, Rome, Stockholm, Vienna",
-  "(GMT+02:00) Athens, Bucharest, Istanbul",
-  "(GMT+03:00) Moscow, St. Petersburg, Volgograd",
-  "(GMT+04:00) Abu Dhabi, Muscat",
-  "(GMT+05:00) Islamabad, Karachi, Tashkent",
-  "(GMT+06:00) Almaty, Novosibirsk",
-  "(GMT+07:00) Bangkok, Hanoi, Jakarta",
-  "(GMT+08:00) Beijing, Perth, Singapore, Hong Kong",
-  "(GMT+09:00) Tokyo, Seoul, Osaka, Sapporo, Yakutsk",
-  "(GMT+10:00) Brisbane, Canberra, Melbourne, Sydney",
-  "(GMT+11:00) Magadan, Solomon Is., New Caledonia",
-  "(GMT+12:00) Auckland, Wellington",
-  "(GMT+13:00) Nuku'alofa"
-];
+export const defaultSteps = [{
+  sequenceOrder: 0,
+  delayDays: 0,
+  delayHours: 0,
+  templateId: 0,
+  campaignId: 0,
+  emailSubject: "",
+  emailBody: "",
+  condition: CampaignEventContition.ALWAYS,
+}];
